fix: video player custom props
